pyproject.toml
[project]
name = "face-emotion-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache==0.12.3",
    "black==24.8.0",
    "fastapi==0.115.0",
    "httpx==0.27.2",
    "isort==5.13.2",
    "msgpack==1.1.0",
    "numpy==1.26.4",
    "passlib[bcrypt]==1.7.4",
    "pillow==10.4.0",
    "prometheus-client==0.21.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "slowapi==0.1.9",
    "structlog==24.4.0",
    "uvicorn[standard]==0.32.0",
    "websocket>=0.2.1",
    "websockets==13.1",
]
